requirements.txt
behave
allure-behave
requests
pyhamcrest
jsonpath-ng
validate_email
